refactor(Purchase): move dir
