+++ server/src/main/java/com/project/legendsofleague/domain/purchase/PurchaseController.java
@@ -1,6 +1,6 @@
-package com.project.legendsofleague.purchase;
+package com.project.legendsofleague.domain.purchase;
 
-import com.project.legendsofleague.purchase.dto.KakaoReadyResponseDto;
+import com.project.legendsofleague.domain.purchase.dto.KakaoReadyResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -31,7 +31,7 @@ public class PurchaseController {
     @Operation(summary = "kakaoPay Approve API.")
     @GetMapping("/purchase/approve")
     public ResponseEntity<String> success(@RequestParam("pg_token") String pgToken,
-                                          @RequestParam(value = "tid", required = false) String tid) {
+        @RequestParam(value = "tid", required = false) String tid) {
 
         String purchaseId = purchaseService.kakaoPaySuccess(pgToken, tid);
         return new ResponseEntity<String>(purchaseId, HttpStatus.OK);
